requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Django==4.2.3
django-compat==1.0.15
django-environ==0.10.0
django_extensions==3.2.3
django-redis==3.1.6
django-redis-cache==3.0.1
djangorestframework==3.14.0
drf-yasg==1.21.6
executing==1.2.0
factory-boy==3.2.1
idna==3.4
importlib-metadata==6.1.0
inflection==0.5.1
ipdb==0.13.13
ipython==8.12.0
jedi==0.18.2
kombu==5.2.4
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytz==2023.3
PyYAML==6.0
redis==3.5.3
redis-simple-cache==0.0.8
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
